extract TankController and ProjectileController from game controller
