+++ pom.xml
@@ -68,6 +68,12 @@
       <artifactId>flexmark</artifactId>
       <version>0.42.14</version>
     </dependency>
+      <dependency>
+          <groupId>org.jetbrains</groupId>
+          <artifactId>annotations</artifactId>
+          <version>23.0.0</version>
+          <scope>compile</scope>
+      </dependency>
   </dependencies>
 
   <build>
