+++ src/main/java/com/example/tankmark1/controllers/MapController.java
@@ -1,4 +1,4 @@
-package com.example.tankmark1;
+package com.example.tankmark1.controllers;
 
 import com.example.tankmark1.map.*;
 import javafx.scene.layout.Pane;
